chore: convert to using python-ffmpeg
