+++ TODO.md
@@ -8,7 +8,7 @@
   - Not really. Size is mostly determined by included codecs.
 - [x] Fix error output so that it's printed if ffmpeg exits on error.
 - [x] Fix to allow modification of video-only file.
-- [ ] Move to python-ffmpeg (ffmpeg-python seems abandoned)
+- [x] Move to python-ffmpeg (ffmpeg-python seems abandoned)
 - [ ] Add countdown timer to progress output?
 - [ ] Default to AV1 encoding instead of H264?
 - [ ] Consider GUI?
