--- snap/snapcraft.yaml.core20.bak
@@ -1,239 +1,0 @@
-name: squeeze-vid
-summary: Normalize videos to a minimal, projection-compatible quality.
-website: https://github.com/sil-car/squeeze-vid
-contact: https://github.com/sil-car/squeeze-vid/issues
-issues: https://github.com/sil-car/squeeze-vid/issues
-license: GPL-3.0+
-title: Squeeze Vid
-icon: pkg-data/video-editor_512.png
-description: |
-  CLI tool to normalize videos and audio files (using ffmpeg) to a standardized
-  frame rate and quality to minimize the file size while maintaining adequate
-  quality for video projection. Options also include changing video playback
-  speed, trimming the video to only keep a portion of it, and exporting the audio
-  from a video file.
-
-  Use the --help option for more information:
-  $ squeeze-vid --help
-
-adopt-info: squeeze-vid # for setting version and grade
-confinement: strict
-base: core20
-architectures:
-  - build-on: [arm64, amd64]
-
-apps:
-  squeeze-vid:
-    command: bin/squeeze-vid
-    # YAML aliases (e.g. "&allplugs"):
-    # https://www.linode.com/docs/guides/yaml-anchors-aliases-overrides-extensions/
-    plugs: &allplugs
-      - home
-      - removable-media
-  ffmpeg:
-    command: usr/bin/ffmpeg
-    plugs: *allplugs
-  ffprobe:
-    command: usr/bin/ffprobe
-    plugs: *allplugs
-
-parts:
-  squeeze-vid:
-    after:
-      - ffmpeg
-    plugin: python
-    source: https://github.com/sil-car/squeeze-vid.git
-    source-depth: 1
-    override-pull: |
-      snapcraftctl pull
-      snap_ver=$(grep 'VERSION =' $SNAPCRAFT_PART_SRC/squeeze_vid/config.py | grep -Eo '[0-9.]+')
-      ffmpeg_ver="ffmpeg7.0" # needs to match parts > ffmpeg > source-branch below
-      snapcraftctl set-version "${snap_ver}+${ffmpeg_ver}"
-      snapcraftctl set-grade "stable"
-    build-packages:
-      - python3-pip
-    override-build: |
-      # Install dephell to convert pyproject.toml to setup.py, which is req'd. by snapcraftctl.
-
-      # core20
-      # https://forum.snapcraft.io/t/building-a-core20-python-snap-using-pyproject-toml/22028/2
-      pip3 install --user dephell[full]
-      $HOME/.local/bin/dephell deps convert --from-path=pyproject.toml --from-format=poetry --to-path=setup.py --to-format=setuppy
-
-      snapcraftctl build
-
-  # Build ffmpeg from source.
-  # https://forum.snapcraft.io/t/proper-way-to-leverage-another-snap-in-a-snap/3893/2
-  ffmpeg:
-    after:
-      - libsvtav1
-    build-packages:
-      - git
-      - g++
-      - make
-      - autoconf
-      - automake
-      - build-essential
-      - cmake
-      - libaom-dev # libaom-av1 encoder is way too slow on 20.04, but its decoder is better
-      - libass-dev
-      - libfreetype6-dev
-      - libmp3lame-dev
-      # - libopus-dev
-      - libsdl1.2-dev
-      # - libsvtav1-dev # 22.04+; or build from src
-      # - libtheora-dev
-      - libtool
-      - libva-dev
-      - libvdpau-dev
-      # - libvorbis-dev
-      - libvpx-dev
-      - libx264-dev
-      # - libx265-dev
-      - libxcb1-dev
-      - libxcb-shm0-dev
-      - libxcb-xfixes0-dev
-      - pkg-config
-      - texinfo
-      - yasm
-      - zlib1g-dev
-    plugin: autotools
-    autotools-configure-parameters:
-      # https://github.com/Archiv8/ffmpeg/blob/main/FFMMPEG-CONFIGURE.md
-      # https://ffmpeg.org/general.html#Supported-File-Formats_002c-Codecs-or-Features
-      - --prefix=/usr
-      - --disable-debug
-      - --disable-doc # removes all docs --disable-{htmlpages,manpages,txtpages}
-      - --disable-ffplay
-      - --disable-static
-      - --enable-gpl
-      - --enable-libaom         # AV1 video encoder
-      - --disable-encoder=libaom
-      - --enable-libass         # subtitles support
-      - --enable-libmp3lame
-      - --enable-libsvtav1      # SVT-AV1 video encoder; requires 22.04+ packages libsvtav1{,-dev}; or build from src
-      - --enable-libvpx         # VP9 video encoder
-      - --enable-libx264        # H.264 video encoder
-      # - --enable-nonfree # makes binary unredistributable; ref: ffmpeg/LICENSE.md
-      - --enable-opengl
-      - --enable-shared
-      - --enable-version3
-    source: git://source.ffmpeg.org/ffmpeg.git
-    source-type: git
-    source-depth: 1 # faster clone?
-    source-branch: 'n7.0'
-    stage-packages:
-      # Ref: snapcraftctl build output.
-      - freeglut3
-      - libasound2
-      - libaom0 # encoder way too slow on 20.04, but decoder is still useful
-      - libass9
-      - libcairo2
-      - libfontconfig1
-      # - libfreetype6
-      - libfribidi0
-      - libglu1-mesa
-      - libgraphite2-3
-      - libharfbuzz0b
-      - libmp3lame0
-      - libnuma1
-      - libogg0
-      # - libopus0
-      - libpixman-1-0
-      - libpng16-16
-      # - libsvtav1 # 22.04+; or build from src
-      # - libtheora0
-      - libva-drm2
-      - libva-x11-2
-      - libva2
-      - libvdpau1
-      # - libvorbis0a
-      - libvorbisenc2
-      - libvpx6
-      - libx264-155
-      # - libx265-179
-      - libxcb-render0
-      - libxcb-shape0
-      - libxcb1
-      - libxrender1
-
-      # External libraries providing hardware acceleration:
-      # v4l2_m2m                vaapi                   vdpau
-
-      # Ref: https://github.com/snapcrafters/obs-studio/blob/master/snap/snapcraft.yaml
-      # - libass9
-      # - libdrm2
-      # - libfdk-aac1
-      # - libmp3lame0
-      # - libopenal1
-      # - libopenjp2-7
-      # - libopus0
-      # - libpulse0
-      # - librsvg2-2
-      # - libsdl2-2.0-0
-      # - libspeex1
-      # - libssl1.1
-      # - libtheora0
-      # - libtwolame0
-      # - libv4l-0
-      # - libv4l2rds0
-      # - libva-drm2
-      # - libva-glx2
-      # - libva-wayland2
-      # - libvdpau-va-gl1
-      # - libvorbis0a
-      # - libvorbisenc2
-      # - libvpx6
-      # - libwebp6
-      # - libwebpmux3
-      # - libx11-6
-      # - libx265-179
-      # - libxau6
-      # - libxcb-shape0
-      # - libxcb-shm0
-      # - libxcb-xfixes0
-      # - libxcb1
-      # - libxdmcp6
-      # - libxext6
-      # - libxml2
-      # - libxv1
-      # - libxvidcore4
-      # - mesa-vdpau-drivers
-      # - ocl-icd-libopencl1
-      # # Only available for amd64 and i386
-      # - try:
-      #   - i965-va-driver
-      #   - mesa-va-drivers
-
-  # Build libsvtav1 from source for core20 or older.
-  # Ref:
-  #   - https://gitlab.com/AOMediaCodec/SVT-AV1/-/blob/master/Docs/Build-Guide.md
-  #   - https://snapcraft.io/docs/cmake-plugin#heading--core20
-  libsvtav1:
-    source: https://gitlab.com/AOMediaCodec/SVT-AV1.git
-    source-type: git
-    source-depth: 1
-    plugin: cmake
-    cmake-parameters:
-      - -G"Unix Makefiles"
-      - -DBUILD_APPS=False
-      - -DCMAKE_BUILD_TYPE=Release
-      - -DCMAKE_INSTALL_PREFIX=/usr
-
-  cleanup:
-    # https://forum.snapcraft.io/t/reduce-size-of-qt5-app-snap/31030/7
-    after:
-      - ffmpeg
-      - libsvtav1
-      - squeeze-vid
-    plugin: nil
-    build-snaps:
-      - core20
-    override-prime: |
-      set -eux
-      cd "/snap/core20/current" && find . -type f,l -exec rm -f "$SNAPCRAFT_PRIME/{}" \;
-      for x in include lib64 usr/include usr/lib/pkgconfig usr/share/doc-base usr/share/ffmpeg/examples usr/share/man; do
-        rm -rf "$SNAPCRAFT_PRIME/$x"
-      done
-      find $SNAPCRAFT_PRIME/usr/share/doc/ -type f -not -name 'copyright' -delete
-      find $SNAPCRAFT_PRIME/usr/share -type d -empty -delete
