+++ squeeze_vid/config.py
@@ -1,4 +1,4 @@
 DEBUG = False
 VERBOSE = False
-VERSION = '0.9.2'
+VERSION = '1.0.0'
 FFMPEG_EXPERIMENTAL = False
