+++ squeeze_vid/util.py
@@ -1,9 +1,6 @@
-import ffmpeg
 import sys
+from ffmpeg import FFmpegError, Progress
 from pathlib import Path
-# https://docs.python.org/3/library/queue.html?highlight=queue#module-queue
-from queue import Queue
-from threading import Thread
 
 from . import config
 
@@ -34,7 +31,7 @@ def parse_timestamp(timestamp):
 
 
 def print_command(stream):
-    command = ffmpeg.get_args(stream)
+    command = stream.arguments[1:]  # omit 'ffmpeg'
     # Add quotes around iffy command arg. options.
     for i, item in enumerate(command.copy()):
         if item == '-filter_complex' or item == '-i':
@@ -49,12 +46,11 @@ def run_conversion(output_stream, duration):
     duration = float(duration)
     if config.DEBUG:
         print(f"{duration=}")
-    filepath = output_stream.node.kwargs.get('filename')
-    print(filepath)
+    print(output_stream.arguments[-1])
 
     def get_progressbar(p_pct, w=60, suffix=''):
         suffix = f" {int(p_pct):>3}%"
-        end = '\n' if config.VERBOSE else '\r'
+        end = '\n' if config.VERBOSE or config.DEBUG else '\r'
 
         ci = '\u23b8'
         cf = '\u23b9'
@@ -78,97 +74,16 @@ def run_conversion(output_stream, duration):
         bar += suffix + end
         return bar
 
-    def read_output(pipe, q):
-        for line in iter(pipe.readline, b''):
-            text = line.decode('utf8')
-            q.put(text)
-        pipe.close()
+    @output_stream.on('progress')
+    def on_progress(progress: Progress):
+        percent = progress.time.total_seconds() * 100 / duration
+        if config.DEBUG or config.VERBOSE:
+            print(progress)
+        sys.stdout.write(get_progressbar(percent))
 
-    def write_output(duration, q):
-        for text in iter(q.get, None):
-            tokens = text.rstrip().split('=')
-            if config.VERBOSE:
-                sys.stdout.write(text)
-            elif len(tokens) == 1 and text[:4] != 'Svt[':  # ignore extra libsvtav1 output  # noqa: E501
-                sys.stdout.write(text)
-            elif len(tokens) == 2:
-                k, v = tokens
-                if k == 'out_time_ms':
-                    try:
-                        current = float(v) / 1_000_000  # convert to sec
-                    except ValueError:
-                        current = 0
-                    # BUG: out_time_ms is initially a very large neg. number in
-                    # ffmpeg6 output; hacked workaround.
-                    if current < 0:
-                        current = 0
-                    if config.DEBUG:
-                        print(f"{current=}")
-                    p_pct = int(round(current * 100 / duration, 0))
-                    progressbar = get_progressbar(p_pct)
-                    sys.stdout.write(progressbar)
-                if config.VERBOSE:
-                    # Only print most interesting progress attributes.
-                    attribs = [
-                        'frame',
-                        'fps',
-                        'total_size',
-                        'out_time',
-                        'speed',
-                    ]
-                    if k in attribs:
-                        sys.stdout.write(text)
-            q.task_done()
-        q.task_done()
-
-    def cleanup(threads, q):
-        for t in threads:
-            t.join()
-        q.put(None)
-        q.join()
-
-    # Run ffmpeg command.
-    if config.DEBUG:
-        ffmpeg.run(output_stream, overwrite_output=True)
-    else:
-        try:
-            with ffmpeg.run_async(
-                output_stream,
-                overwrite_output=True,
-                pipe_stdout=True,
-                pipe_stderr=True,
-            ) as p:
-                q_out = Queue()
-                try:
-                    # Start progress queue & threads.
-                    t_out = Thread(
-                        name="T-stdout",
-                        target=read_output,
-                        args=(p.stdout, q_out)
-                    )
-                    t_err = Thread(
-                        name="T-stderr",
-                        target=read_output,
-                        args=(p.stderr, q_out)
-                    )
-                    t_write = Thread(
-                        name="T-write",
-                        target=write_output,
-                        args=(duration, q_out)
-                    )
-                    for t in (t_out, t_err, t_write):
-                        t.start()
-                    p.wait()
-                except KeyboardInterrupt:
-                    print("\nInterrupted with Ctrl+C")
-                    p.kill()
-                    # Finish queue & progress bar.
-                    cleanup((t_out, t_err), q_out)
-                    sys.exit(130)
-                # Finish queue & progress bar.
-                cleanup((t_out, t_err), q_out)
-        except ffmpeg._run.Error as e:
-            print(e.stderr.decode('utf8'))
-            exit(1)
-
-    print()
+    try:
+        output_stream.execute()
+        sys.stdout.write(get_progressbar(100)) # for a nice, cleann finish
+        print()
+    except FFmpegError as e:
+        print(f"{e.message}: {e.arguments}")
